Show Average instead of Total in access distrib
In the access distribution, we used to show the total number of access per
pages. As the total is often bigger than the number of access per thread, it
couldn't be displayed in the same scale than the number of access per thread.
This lead to a biased representation. Using the average instead of the total
fix that issue.

This commit also add a bit more of color for first touch plots, and a cleaner
color visualization for the other plots.